Support for register to vote

+++ generator/src/main/scala/io/michaelallen/mustache/generator/MustacheGenerator.scala
@@ -3,15 +3,19 @@ package io.michaelallen.mustache.generator
 import java.io.File
 import sbt.IO
 import sbt.{PathExtra, FileFilter}
+import io.michaelallen.logging.Timing
 
-trait MustacheGenerator extends PathExtra {
+trait MustacheGenerator extends PathExtra with Timing {
 
-  def writeFile(file:File, content:String) = IO.write(file, content)
+  def writeFile(file:File, content:String) = time(s"writing to ${file.name}") {
+    IO.write(file, content)
+  }
 
   def templateContent(
       namespace: Seq[String],
       name: String,
-      template: String
+      template: String,
+      templateHash: String
   ): String = {
     s"""
       |package ${namespace.mkString(".")}
@@ -20,6 +24,7 @@ trait MustacheGenerator extends PathExtra {
       |import io.michaelallen.mustache.MustacheFactory
       |
       |object $name {
+      |  val hash: String = "$templateHash"
       |  val mustache: Mustache = MustacheFactory.compile("$template")
       |}
       |
@@ -69,8 +74,10 @@ trait MustacheGenerator extends PathExtra {
 
   def generatePlaySource(sourceTarget: File): File = {
     val file = sourceTarget / "io" / "michaelallen" / "mustache" / "PlayImplicits.scala"
-    val content = playImplicitsContent()
-    writeFile(file, content)
+    if (!file.exists) {
+      val content = playImplicitsContent()
+      writeFile(file, content)
+    }
     file
   }
 
@@ -84,19 +91,22 @@ trait MustacheGenerator extends PathExtra {
     val excluded = directory ** excludeFilter
     val templates = included --- excluded
     val mappings = templates pair relativeTo(directory)
-    mappings map { case (file, path) =>
+    mappings.foldLeft(Seq.empty[File]) { case (files, (file, path)) =>
       val name = file.base
       val ext = file.ext
       val template = path
+      val templateHash = file.hashString
 
       val parentPath = Option(new File(path).getParent)
       val namespace = "mustache" +: parentPath.map(_.split("/")).toSeq.flatten
       val relativePath = namespace.mkString("/")
 
       val targetFile = sourceTarget / relativePath / s"$name.scala"
-      val content = templateContent(namespace, name, template)
-      writeFile(targetFile, content)
-      targetFile
+      if (!targetFile.exists || targetFile.olderThan(file)) {
+        val content = templateContent(namespace, name, template, templateHash)
+        writeFile(targetFile, content)
+      }
+      files :+ targetFile
     }
   }
 
@@ -107,7 +117,7 @@ trait MustacheGenerator extends PathExtra {
       includeFilter: FileFilter,
       excludeFilter: FileFilter,
       createPlayImplicits: Boolean
-  ): Seq[File] = {
+  ): Seq[File] = time("Generate Scala sources") {
     val sourcesInSeqs = sourceDirectories map { directory =>
       generateTemplateSourcesForDirectory(
         directory, sourceTarget, includeFilter, excludeFilter
@@ -130,8 +140,10 @@ trait MustacheGenerator extends PathExtra {
       mustacheTarget: String
   ): File = {
     val file = targetDir / "io" / "michaelallen" / "mustache" / "MustacheFactory.scala"
-    val content = factoryObjectContent(mustacheTarget)
-    writeFile(file, content)
+    if (!file.exists) {
+      val content = factoryObjectContent(mustacheTarget)
+      writeFile(file, content)
+    }
     file
   }
 }
